puntero(pocho)

+++ MonkeyDelivery/Src/Items/Banana.cpp
@@ -11,7 +11,7 @@ Banana::Banana(Texture* tex, Game* game, Player* player) : InventoryObject(tex, 
 bool Banana::useObject()
 {
 	player_->initPowerUp(banana);
-	sdlutils().soundEffects().at("soda").setVolume(game_->getSoundEfectsVolume() * game_->getGeneralVolume());
+	sdlutils().soundEffects().at("soda").setVolume((int)(game_->getSoundEfectsVolume() * game_->getGeneralVolume()));
 	sdlutils().soundEffects().at("soda").play(0, 1);
 	return true;
 }
